requirements.txt
Flask
Flask-SocketIO
eventlet
geopy
gunicorn
python-dotenv
google-generativeai
